// src/features/exercise/filter-exercise/model/types.ts
export interface IFilterOptions {
  skill: (string | number)[];
  type: (string | number)[];
  studentLevel: string;
  studentAge: string;
  topicList: (string | number)[];
}

// export interface IFilterOptions {
//   skill: (string | number)[];
//   type: (string | number)[];
//   studentLevel: (string | number)[];
//   studentAge: (string | number)[];
//   topicList: (string | number)[];
// }